Extract constructor options as interface
No functional change, only extracting the constructor options as
an exported interface to simplify for clients that need a type to
refer to.
For example, a client wanting to pass a subset of options currently
would have to use constructs such as
`Pick<ConstructorParameters<typeof OpenAPI.OpenAPIBackend>[0], 'ajvOptions'>`,
but can now use `Pick<OpenAPI.Options, 'ajvOptions'>`.